Room-change travel cost: a sense outside the robot's room costs 1 + c
Every active sense used to cost one budget unit wherever the robot stood.
It now costs 1.0 for a receptacle in the room the robot is already in and
1 + c for one anywhere else. One parameter, distance-free, and at c = 0
exactly the model the package had before. What a sense REVEALS is
unchanged, and the passive patrol is unchanged.

Bank headers gain `receptacle_rooms` and `home_base_room` (optional; the
loader rejects one without the other). Episode/EpisodeContext carry them,
plus `EpisodeContext.sense_cost()` and a live RobotPosition the harness
owns, so policies price their options without widening `decide`.
`run_episode(agent, episode, room_change_cost=0.0)` tracks position (home
base at each day start, the room of every ambient visit delivered up to
t_query, the room of every receptacle sensed), charges cost(r), and
refuses a sense it cannot afford. Budgets are floats throughout; the
per-question loop is bounded by the sensable-receptacle count, which
terminates under fractional costs where the old budget-derived bound
would not. Only VoIThresholdSense and VoIBudgetPriceSense can see the
price: they sense while voi(r) >= lambda * cost(r) and pick the
receptacle maximizing voi(r) / cost(r). Every other policy is untouched.

New: `room_change_cost_study.py` (driver), `bank.write_room_cost_bank`
(fixture), `tests/test_baselines_room_cost.py` (29 tests). mypy --strict
clean on every new and touched file. The golden run log is regenerated
for the schema change; note it did not reproduce at HEAD on this machine
beforehand either (one ULP in a distribution), which predates this work.

Results in `results/room_change_cost/` over 20 banks re-exported with
room maps (byte identical to the old ones but for the two header fields;
`bank_verification.md`), 4 beliefs x 5 policies x 2 budgets x 5 costs x
10 test households.

The c = 0 regression against results/voi_policies/ fails on 10 of 30
cells, and the cause is pre-existing: that csv was generated on a dirty
tree 3.6 h before the Part B commit, with a different voi_sense.py, so no
committed code reproduces it. Re-running the mismatching cells under
unmodified HEAD reproduces THIS study's numbers exactly on all 7 tested
(head_baseline.json), so the cost model changed nothing at c = 0.
results/voi_policies/'s VoI rows are stale by up to 0.008 and should be
regenerated before they are quoted again.

Findings are marked PROVISIONAL: the query distribution is being redrawn,
and the arrival result (the robot's room is no more informative about the
answer than independence gives) is a property of the patrol crossed with
that distribution. A high-cost compensated-budget arm is implemented
behind --costs/--compensate-budget and deliberately not run for the same
reason.

+++ src/baselines/agent.py
@@ -43,7 +43,7 @@ class Agent:
         return self.belief.predict(question.object_id, question.t_query)
 
     def decide(self, question: Question, prediction: Prediction,
-               budget_remaining: int,
+               budget_remaining: float,
                last_sense: Union[SenseResult, None] = None) -> Action:
         """Delegate the sense-or-answer choice to the policy."""
         return self.policy.decide(question, prediction, budget_remaining,
