+++ src/baselines/policies/aci_sense.py
@@ -119,7 +119,7 @@ class ACISense(DecisionPolicy):
         self.n_updates += 1
 
     def decide(self, question: Question, prediction: Prediction,
-               budget_remaining: int, t: int,
+               budget_remaining: float, t: int,
                last_sense: Optional[SenseResult] = None) -> Action:
         qhat = self.qhat()
         members = prediction_set(prediction.distribution, qhat,
